JPERF-917 Upgrade Ubuntu image to 20.04 LTS. This will unblock migration to IAM in aws-infrastructure lib.
When using the old image I encountered problems with Google Chrome installation caused by missing dependency, with newer image the problem is gone.
Initial goal was to upgrade to the newest Ubuntu (22.04), but on this version I encountered problems with ssh so I decided to tackle this issue in a separate task.