JPERF-1332: Reduce pressure on CloudFormation
When there's a lot of expired stacks (e.g. due to housekeeping outage),
then listing them all requires a lot of Cloudformation requests.
Then we get throttled, so AWS SDK makes a couple retries underneath,
but we finally time out. Then we don't reach the stack deletion part.
So the list remains long and it's a self-perpetuating problem.

Instead, start deleting stacks as soon as we get a batch listed.